unstable: map parserCreation page msg to 'application' msg
The method for viewing the create parser page now produces 'Html
ParserCreation.Msg'. To deal with it in the Main.elm file it is mapped
to another type of message. This will allow me to move easily move the
entire view and message code to another file (see also: richard
feldman spa).
This commit throws a runtime exception (Debug.todo).